% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_methods.R
\name{print.summary.robust_gmm}
\alias{print.summary.robust_gmm}
\title{Print a Summarized Robust Growth Mixture Model}
\usage{
\method{print}{summary.robust_gmm}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.robust_gmm"}.}

\item{digits}{Number of decimal places. Default \code{3}.}

\item{...}{Currently ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Summarized Robust Growth Mixture Model
}
\keyword{internal}
