% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_mean.R
\name{robust_mean}
\alias{robust_mean}
\title{Calculate the robust mean}
\usage{
robust_mean(data, threshold = 10)
}
\arguments{
\item{data}{A matrix or data.frame}

\item{threshold}{Maximum distance allowed to not be considered an outlier}
}
\value{
A numeric vector representing the robust mean of the variables.
}
\description{
Calculate the robust mean
}
\examples{
data(iris)
r_mean <- robust_mean(iris[1:30, 1:2])

# Print the calculated robust means
r_mean
}
