% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustLPA-package.R
\docType{package}
\name{RobustLPA-package}
\alias{RobustLPA}
\alias{RobustLPA-package}
\title{RobustLPA: Robust Latent Profile Analysis}
\description{
Provides a comprehensive toolset for estimating Latent Profile Analysis (LPA) models that are robust to multivariate outliers and missing data. By integrating a high-performance 'C++' engine via 'RcppArmadillo', it reliably extracts latent profiles using both Expectation-Maximization (EM) and Markov Chain Monte Carlo (MCMC) Bayesian estimation. The EM engine implements a Full Information Maximum Likelihood (FIML) approach, Huber weighting, and LASSO regularization with k-fold cross-validation for optimal penalty tuning. The MCMC engine utilizes a Bayesian Lasso approach with Laplace priors, the same Huber down-weighting available in the EM engine, multiple chains (4 by default), and classic Gelman-Rubin/effective sample size convergence diagnostics. It supports multiple geometric variance-covariance models, along with functions for bootstrapped likelihood ratio tests (BLRT), BCH auxiliary variable analysis, and plotting. For methodological details on the Bootstrapped Likelihood Ratio Test, see Nylund et al. (2007) \doi{10.1080/10705510701575396}. For robust clustering methods, see Garcia-Escudero et al. (2010) \doi{10.1007/s11634-010-0064-5}. For BCH auxiliary variable analysis, see Bolck et al. (2004) \doi{10.1093/pan/mph001}.
}
\author{
\strong{Maintainer}: Valerio Riccardo Aquila \email{valerio_aquila@hotmail.it} (\href{https://orcid.org/0009-0004-2231-2141}{ORCID})

Authors:
\itemize{
  \item Valerio Riccardo Aquila \email{valerio_aquila@hotmail.it} (\href{https://orcid.org/0009-0004-2231-2141}{ORCID})
}

}
\keyword{internal}
