% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_robust_lpa.R
\name{plot_robust_lpa}
\alias{plot_robust_lpa}
\title{Plot Robust Latent Profiles}
\usage{
plot_robust_lpa(
  model,
  title = "Robust Latent Profiles",
  xlab = "Variables",
  ylab = "Value",
  var_labels = NULL,
  legend_title = "Class"
)
}
\arguments{
\item{model}{A model object returned by robust_lpa or estimate_profiles_robust.}

\item{title}{The title of the plot. Default is "Robust Latent Profiles".}

\item{xlab}{The x-axis label. Default is "Variables".}

\item{ylab}{The y-axis label. Default is "Value".}

\item{var_labels}{A character vector to manually rename the variables on the X axis. Default is NULL (auto-detect).}

\item{legend_title}{The title of the legend. Default is "Class".}
}
\value{
A ggplot object.
}
\description{
Automatically generates a professional profile plot using ggplot2 from an estimated
robust LPA model.
}
\examples{
data(iris)
fit <- robust_lpa(data = iris[1:30, 1:2], G = 2, model = 1, n_starts = 1)
plot_robust_lpa(fit)
}
